refactor(article): define base package as "nettee.blolet.article.domain" in article domain module
