+++ services/article/api/domain/src/main/java/nettee/blolet/article/domain/sub/ArticleStatus.java
@@ -1,4 +1,4 @@
-package nettee.article.domain;
+package nettee.blolet.article.domain.sub;
 
 public enum ArticleStatus {
     ACTIVE,
