+++ services/article/driven/rdb/src/main/java/nettee/article/driven/rdb/entity/type/builder/ArticleLikesEntityStatus.java
@@ -1,7 +1,7 @@
 package nettee.article.driven.rdb.entity.type.builder;
 
-import nettee.article.domain.ArticleLikesStatus;
-import nettee.article.domain.ArticleStatus;
+import nettee.blolet.article.domain.sub.ArticleLikesStatus;
+import nettee.blolet.article.domain.sub.ArticleStatus;
 import nettee.common.marker.TypeSafeMarker.Present;
 import nettee.common.status.StatusCodeUtil;
 import nettee.common.status.StatusParameters;
@@ -10,7 +10,7 @@ import java.util.Arrays;
 import java.util.Set;
 import java.util.stream.Collectors;
 
-import static nettee.article.exception.ArticleErrorCode.DEFAULT;
+import static nettee.blolet.article.exception.ArticleErrorCode.DEFAULT;
 import static nettee.common.status.StatusParameters.GeneralPurposeFeatures.ALL;
 import static nettee.common.status.StatusParameters.GeneralPurposeFeatures.SUBITEM_READ;
